+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-# v1.0.0
+# v2.0.0
 
-Document features and fixes
+- Help in reactangle terrain mode only shown in create mode
+- Add units to ui components
+- Update core and ui to 6.0